✨ Migrate ProfilePage component to new structure with token-based authentication and localStorage userId retrieval
